SEM6/sourcecode/backend/requirements.txt
Flask==3.0.3
flask_cors==5.0.0
Pillow==10.4.0
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
ultralytics==8.3.11
Werkzeug==3.0.4
